# Cleanups and optimization of ToBig()