```
name                     old time/op  new time/op  delta
ToBig/1word-6            64.9ns ± 0%  50.0ns ± 1%  -22.90%  (p=0.008 n=5+5)
ToBig/2words-6           90.6ns ± 0%  50.0ns ± 0%  -44.86%  (p=0.008 n=5+5)
ToBig/3words-6            111ns ± 0%    50ns ± 2%  -55.14%  (p=0.008 n=5+5)
ToBig/4words-6            126ns ± 0%    50ns ± 1%  -60.59%  (p=0.008 n=5+5)
_MulMod/large/uint256-6   628ns ± 0%   531ns ± 0%  -15.45%  (p=0.016 n=5+4)
```

1. Merge FromBig() and NewFromBig(). They were identical.
2. Adds more unit tests. Moves conversion tests to `conversion_test.go`.
3. Adds a benchmark for `ToBig()`.
4. Optimizes `ToBig()` in a similar way as `FromBig()` although in this case implementation is much simpler.

Probably the best to review commit by commit.